refactor: view 리팩토링 및 패키지 수정
