+++ src/main/java/racing/domain/RandomMovableRule.java
@@ -1,4 +1,4 @@
-package racing.car;
+package racing.domain;
 
 import racing.util.RandomUtil;
