+++ src/main/java/racing/view/InputView.java
@@ -1,0 +1,23 @@
+package racing.view;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Scanner;
+
+public class InputView {
+
+    private static Scanner scanner = new Scanner(System.in);
+
+    public static List<String> getCarNames() {
+        System.out.println("경주할 자동차 이름을 입력하세요. (이름은 쉼표(,)를 기준으로 구분)");
+        String inputNames = scanner.nextLine();
+
+        return Arrays.asList(inputNames.split(","));
+    }
+
+    public static int getTimes() {
+        System.out.println("시도할 회수는 몇 회 인가요?");
+        return scanner.nextInt();
+    }
+
+}
