+++ src/main/java/racing/view/ResultView.java
@@ -1,7 +1,7 @@
-package racing.ui;
+package racing.view;
 
-import racing.car.CarRacing;
-import racing.car.model.RaceRecord;
+import racing.domain.CarRacing;
+import racing.domain.RaceRecord;
 import racing.util.StringUtil;
 
 import java.util.List;
@@ -10,13 +10,13 @@ public class ResultView {
 
     public static final String RAIL_STRING = "_";
 
-    public void printResult(CarRacing carRacing) {
+    public static void printResult(CarRacing carRacing) {
         List<RaceRecord> raceRecords = carRacing.getRaceRecords();
-        raceRecords.forEach(this::printRaceRecord);
+        raceRecords.forEach(ResultView::printRaceRecord);
         printWinner(carRacing);
     }
 
-    private void printRaceRecord(RaceRecord raceRecord) {
+    private static void printRaceRecord(RaceRecord raceRecord) {
         System.out.println("RACE :: " + raceRecord.getRaceId());
         raceRecord.getTrackRecords().forEach(
             (key, value) -> System.out.printf("%s : %s \n", key, StringUtil.makeRepeatString(value, RAIL_STRING))
@@ -24,7 +24,7 @@ public class ResultView {
         System.out.println();
     }
 
-    private void printWinner(CarRacing carRacing) {
+    private static void printWinner(CarRacing carRacing) {
         String winners = String.join(", ", carRacing.findWinnerNames());
         System.out.println(winners + "가 최종 우승 했습니다.");
     }
